Separate About and Partner sections into two different pages
