+++ src/pages/about.js
@@ -9,35 +9,6 @@ import SEO from '../components/seo';
 import styles from './about.module.scss';
 
 const AboutPage = () => {
-  const data = useStaticQuery(graphql`
-    query MyQuer {
-      sponsors: allFile(filter: { relativeDirectory: { eq: "sponsors" } }) {
-        nodes {
-          name
-          id
-          childImageSharp {
-            fluid {
-              ...GatsbyImageSharpFluid
-            }
-          }
-        }
-      }
-    }
-  `);
-
-  // const teamData = data.teamData.edges
-  // const teamPics = data.teamPics.nodes
-  const sponsors = data.sponsors.nodes;
-
-  // const getImgSrc = name =>
-  //   teamPics.filter(image => {
-  //     if (!image.name.includes(name)) {
-  //       return null
-  //     }
-
-  //     return image
-  //   })[0].childImageSharp.fluid
-
   return (
     <Layout className={styles.page}>
       <SEO title="About" />
@@ -97,68 +68,6 @@ const AboutPage = () => {
         </div>
       </section>
 
-      <br />
-      <hr className={cx(styles, 'content')} />
-      <br />
-
-      {/* <section className={cx(styles.gridContainer12, 'content')}>
-        <h1 className={cx(styles, 'content-max-width')}>
-          Meet the Team
-        </h1>
-        <div className={styles.team}>
-
-          {teamData.map(person => (
-            <div key={person.node.id} className={styles.person}>
-              <div className={styles.image}>
-                <Img
-                  className={cx(styles.img)}
-                  fluid={getImgSrc(person.node.name)}
-                  alt={`Image of ${person.node.name}`}
-                />
-              </div>
-              <h4 className={styles.name}>{person.node.name}</h4>
-              <div className={styles.role}>{person.node.role}</div>
-            </div>
-          ))}
-        </div>
-      </section>
-      <br />
-      <br />
-      <br /> */}
-
-      <section className={cx(styles.gridContainer12, 'content')}>
-        <h2 className={cx(styles, 'content-max-width')}>Our Partners</h2>
-        <div>
-          <p>
-            We believe that realizing our vision of creating life-changing
-            impact for our communities through developing exceptional digital
-            experiences and educational programming means partnering with
-            inspiring organizations to catalyze our shared work. These are some
-            of the incredible organizations with which we partner.
-          </p>
-        </div>
-      </section>
-
-      <br />
-
-      <section className={cx(styles.gridContainer12, 'content')}>
-        <div className={styles.partners}>
-          {/* <h1 className={cx(styles.subtitle, 'content-max-width')}>
-            These are some of the incredible organizations we partner with.
-          </h1> */}
-          <div className={styles.logoGroup}>
-            {sponsors.map((logo) => (
-              <div key={logo.id} className={styles.logoItem}>
-                <Img
-                  className={cx(styles.logo)}
-                  fluid={logo.childImageSharp.fluid}
-                  alt={logo.name}
-                />
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
     </Layout>
   );
 };
