+++ src/pages/partners.module.scss
@@ -1,0 +1,41 @@
+@import '../_assets/css/abstracts/1-abstracts';
+/*
+ *
+ * About
+ *
+ */
+
+.team {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(168px, 168px));
+  column-gap: var(--sp-max);
+  row-gap: var(--sp-loose);
+}
+
+.subtitle {
+  margin-top: 100px;
+  margin-bottom: 50px;
+}
+
+.name {
+  font-size: var(--fs-base);
+  text-transform: capitalize;
+  font-family: var(--font-italic);
+  margin-top: var(--sp-tight);
+  margin-bottom: var(--sp-tight);
+}
+
+.partnersTitle {
+  margin-top: var(--sp-loosest);
+  margin-bottom: var(--sp-loose);
+}
+
+.logoGroup {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(auto, 160px));
+  gap: var(--sp-loosest);
+}
+
+.logoItem {
+  align-self: center;
+}
